+++ packages/utils/updater/src/utils/package.test.ts
@@ -1,0 +1,14 @@
+import {
+	describe,
+	it,
+} from 'vitest'
+
+describe( 'Package', () => {
+
+	it( 'Shoud be correct', async () => {
+
+		await import( './package.example' )
+
+	} )
+
+} )